requirements.txt
kafka-python==2.0.2
fastapi[all]== 0.75.1
python-dotenv==0.20.0
Faker==13.3.4